feat: add modules/esm.mjs solution
